requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Talisman==1.1.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
itsdangerous==2.2.0
qrcode==7.4.2
Pillow==10.4.0
reportlab==4.2.0
XlsxWriter==3.2.0
python-dotenv==1.0.1
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
sentry-sdk==2.8.0
pytesseract==0.3.10
twilio==9.2.3
openai==1.35.10


